Bootstrap source macro
When building from git, the initial .tar.gz is built along the procedure. Rpm
doesn't like that, since it requires all Sources to be present at bootstrap.

With this macro, wrapped around the source file, an empty .tar.gz is first
generated during the expansion of the macro, which keeps rpmbuild happy.