Sumbodules: allow forced cleanup of repos, at %prep stage..
If repos are dirty (eg. the previous build phase had touched files),
they need to be cleaned and brought to the checkout state, before
building again from them. So, "%git_prep_submodules -f" at the spec file
will tell us that we need a clean version of our build clones.

Note: we are talking about the cloned (=copied) repos and /not/ the ones
that data came from (= the development ones).